test: adapt tests for SecuredMedia
